Finished Drag n Drop in inventory
Completed code for the drag n drop section of the tutorial. While its not implemented for the spells, one solution could be to follow the hotbar tutorial (video 5 in the playlist) and make it so that you can only put spellclass items in the hotbar and sturcture it in a quad diamond formation.