Inventory System Foundation Complete
This commit updates Part-Time Warlock with the foundation for an inventory system. This allows players to add, remove,  and swap the positions of items, as well as a basic crafting system. In addition, this system also allows the player to sync their four spell slots (and potion slot) with the inventory. This system will be expanded upon to deliver several of the game's core mechanics, such as shopping for spells and a more robust potion making system.