Finished cooldown system
Spells now have a cooldown system, the values of which can be set in the inspector for each spell's respective scriptable object. This system doesn't have a UI representation for it yet, but has complete functionality on the backend side of things.

Currently, this has only been tested with projectile spells, but this cooldown system should theoretically work with the other three spell types (area of effect, movement, and status).