Finished Conversion to Unity 2022 LTS
After removing some package dependencies from Unity 2023, downgrading to Unity 2022 was successful, in the sense that the project didn't break when trying to open it in Unity 2022. There are still some game specific bugs, but the project can be successfully compiled, built, and run.

This commit also includes removing some redundant/unnecessary code, as was explained in a previous commit that removed the previous spell system.