Started New Inventory System
Tried as I might, I could not get the existing inventory system implemented with a chest/shop, which is required for many of the systems we want to create (e.g. the apartment). I had to start from scratch and start creating a new one. The old system will remain in the codebase until the new one is implemented.

This push also contains the test level with the newly implemented art