Further Inventory system progress
The player now has a backpack as well as the game having the beginning of a chest system. The one problem is that, currently, neither of them are saved between runs. Will be fixed in the next push.

I left of at the beginning of this video in the inventory system tutorial:
https://www.youtube.com/watch?v=f1cqInqB9qk&list=PL-hj540P5Q1hLK7NS5fTSNYoNJpPWSL24&index=10