Inventory system update
Removed old inventory system code and replaced it with new code from the UnityAssetStore. New code can be found under Assets/nappin

currently, there is a bug with coins in the dungeon and them not being able to get picked up. Open/close inventory with tab. Click and drag on items to move them places. Ignore the input directions at the top; I changed the code and some of them are no longer accurate.